Add Customer and DataSource classes for bank application

+++ src/main/java/bank/Account.java
@@ -3,10 +3,10 @@ package bank;
 public class Account {
   private int id;
   private String type;
-  private int balance;
+  private double balance;
 
 
-  public Account(int id, String type, int balance) {
+  public Account(int id, String type, double balance) {
     setId(id);
     setType(type);
     setBalance(balance);
@@ -29,11 +29,11 @@ public class Account {
     this.type = type;
   }
 
-  public int getBalance() {
+  public double getBalance() {
     return this.balance;
   }
 
-  public void setBalance(int balance) {
+  public void setBalance(double balance) {
     this.balance = balance;
   }
